requirements.txt
fastapi
uvicorn[standard]
google-generativeai>=0.6.0
python-dotenv
supabase
